week14/requirements.txt
Django==2.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-extensions==0.6.0
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
